+++ packages/typespec-ts/src/index.ts
@@ -572,7 +572,9 @@ export async function $onEmit(context: EmitContext) {
       if (option.azureSdkForJs) {
         commonBuilders.push(buildTsSrcEsmConfig);
         commonBuilders.push(buildTsSrcBrowserConfig);
-        commonBuilders.push(buildTsSrcReactNativeConfig);
+        if (option.generateReactNativeTarget) {
+          commonBuilders.push(buildTsSrcReactNativeConfig);
+        }
         commonBuilders.push(buildTsSrcCjsConfig);
         if (option.generateSample) {
           commonBuilders.push(buildTsSampleConfig);