+++ packages/typespec-ts/src/utils/operationUtil.ts
@@ -193,11 +193,18 @@ export function isDefinedStatusCode(statusCodes: HttpStatusCodesEntry) {
 export function isBinaryPayload(
   dpgContext: SdkContext,
   body: Type,
-  contentType: string | string[]
+  contentType: string | string[],
+  encode?: string
 ) {
   const knownMediaTypes: KnownMediaType[] = (
     Array.isArray(contentType) ? contentType : [contentType]
   ).map((ct) => knownMediaType(ct));
+
+  // When encode is "bytes" treat as binary.
+  if (encode === "bytes") {
+    return true;
+  }
+
   for (const type of knownMediaTypes) {
     if (type === KnownMediaType.Binary && isByteOrByteUnion(dpgContext, body)) {
       return true;